Add comment delete part2 delete permission and Http status code.

+++ mainBlog/comments/views.py
@@ -1,13 +1,25 @@
 from django.contrib.contenttypes.models import ContentType
 from django.contrib import messages
-from django.http import HttpResponseRedirect
+from django.http import HttpResponseRedirect, Http404, HttpResponse
 from django.shortcuts import render, get_object_or_404
 from .forms import CommentForm
 from .models import Comment
 
 
 def comment_delete(request, id):
-    obj = get_object_or_404(Comment, id=id)
+    # obj = get_object_or_404(Comment, id=id)
+    try:
+        obj = Comment.objects.get(id=id)
+    except:
+        raise Http404
+
+    if obj.user != request.user:
+        # messages.success(request, "Yoy do not have permission to view this.")
+        # raise Http404
+        response = HttpResponse("Yoy do not have permission to do this.")
+        response.status_code = 403
+        return response
+
     if request.method == "POST":
         parent_obj_url = obj.content_object.get_absolute_url()
         obj.delete()
@@ -20,7 +32,15 @@ def comment_delete(request, id):
 
 
 def comment_thread(request, id):
-    obj = get_object_or_404(Comment, id=id)
+    # obj = get_object_or_404(Comment, id=id)
+    try:
+        obj = Comment.objects.get(id=id)
+    except:
+        raise Http404
+
+    if not obj.is_parent:
+        obj = obj.parent
+
     content_object = obj.content_object
     content_id = obj.content_object.id
